vendor & application name

+++ JumpstartRealmJoinPackage.ps1
@@ -1,20 +1,24 @@
 Param(
-    [string] $RepositoryPath,
-    [string] $RepositoryName,
-    [string] $RepositoryNamespace,
-    [string] $GitlabPersonalAccessToken = $Env:GitLabToken,
-    [switch] $GitUseSsh,
-    [switch] $DoNotQueryParameters,
-    [switch] $DoNotCreateRepository,
-    [switch] $DoNotCloneRepository,
-    [switch] $DoNotCopyTemplate,
-    [switch] $DoNotRunTemplateScript,
-    [Parameter(ValueFromRemainingArguments = $true)] $RemainingArgumentsToPassToTemplate
+    [string] $vendorName,
+    [string] $applicationName,
+    [string] $repositoryName,
+    [string] $repositoryPath,
+    [string] $repositoryNamespace,
+    [string] $gitlabPersonalAccessToken = $Env:GitLabToken,
+    [switch] $gitUseSsh,
+    [switch] $doNotQueryParameters,
+    [switch] $doNotCreateRepository,
+    [switch] $doNotCloneRepository,
+    [switch] $doNotCopyTemplate,
+    [switch] $doNotRunTemplateScript,
+    [switch] $debugUseLocalTemplates,
+    [Parameter(ValueFromRemainingArguments = $true)] $remainingArgumentsToPassToTemplate
 
 )
 
 $ErrorActionPreference = "Stop"
 Set-StrictMode -Version 2
+Write-Host
 
 
 if (-not $gitlabPersonalAccessToken) {
@@ -26,12 +30,29 @@ if (-not $gitlabPersonalAccessToken) {
 }
 
 if (-not $DoNotQueryParameters) {
-    Write-Host "Querying missing parameters about RealmJoin GitLab:"
-    if (-not $repositoryPath) {
-        $repositoryPath = Read-Host "Repository path (leave empty for current folder name, Format: {vendor}-{productname})"
+    Write-Host "*** Querying missing package parameters ***"
+    Write-Host "* For vendor name, use a short name, e.g. just 'Microsoft' instead of "
+    Write-Host "  'Microsoft Corporation'."
+    Write-Host "* For config packages or packages with very simple names (e.g. open source "
+    Write-Host "  projects like 'Putty' etc.), vendor name may be left empty."
+    Write-Host "* For organic packages, use 'Organic' as vendor name and the full appliction name "
+    Write-Host "  incl. vendor as application name, e.g. 'Organic' and 'Microsoft Coding Toolbox'."
+    Write-Host
+    if (-not $vendorName) {
+        $vendorName = (Read-Host "Vendor name").Trim()
+    }
+    while (-not $applicationName) { # we really need this, so we loop until we get it
+        $applicationName = (Read-Host "Application name").Trim()
     }
     if (-not $repositoryName) {
-        $repositoryName = Read-Host "Repository name (leave empty for repository path)"
+        $repoNameDefault = "$vendorName $applicationName".Trim()
+        $repositoryName = (Read-Host "Repository name (default: '$repoNameDefault')").Trim()
+        if (-not $repositoryName) { $repositoryName = $repoNameDefault }
+    }
+    if (-not $repositoryPath) {
+        $repoPathDefault = ($repositoryName -ireplace '[-_ ]+', '-' -ireplace '[^a-z0-9-]').Trim('-').ToLowerInvariant()
+        $repositoryPath = (Read-Host "Repository path (default: $repoPathDefault)").Trim()
+        if (-not $repositoryPath) { $repositoryPath = $repoPathDefault }
     }
     if (-not $repositoryNamespace) {
         $repositoryNamespace = Read-Host "Repository namespace (leave empty for 'generic-packages', Format: {customer}-packages)"
@@ -42,18 +63,22 @@ if (-not $DoNotQueryParameters) {
     if (-not $PSBoundParameters.ContainsKey("GitUseSsh")) {
         $gitUseSsh = [switch]((Read-Host "Use SSH for Git [y/N] (default is https)") -in "y", "j", "1", "true")
     }
+} else {
+    if (-not ($applicationName)) {
+        throw "ERROR: Need ApplicationName if passing -DoNotQueryParameters"
+    }
+    if (-not $repositoryName) {
+        $repositoryName = "$vendorName $applicationName".Trim()
+    }
+    if (-not $repositoryPath) {
+        $repositoryPath = ($repositoryName -ireplace '[-_ ]+', '-' -ireplace '[^a-z0-9-]').Trim('-').ToLowerInvariant()
+    }
 }
 
-if ($repositoryPath) {
-    New-Item $repositoryPath -Type Directory -ErrorAction SilentlyContinue | Out-Null
-    Set-Location $repositoryPath
-}
-else {
-    $repositoryPath = [System.IO.Path]::GetFileName((Get-Location).Path)
-}
-if (-not $repositoryName) {
-    $repositoryName = $repositoryPath
-}
+
+New-Item $repositoryPath -Type Directory -ErrorAction SilentlyContinue | Out-Null
+Set-Location $repositoryPath
+
 if (-not $repositoryNamespace) {
     $repositoryNamespace = "generic-packages"
 }
@@ -99,6 +124,11 @@ else {
 
 }
 
+Write-Debug "VendorName:          $vendorName"
+Write-Debug "ApplicationName:     $applicationName"
+Write-Debug "RepositoryName:      $repositoryName"
+Write-Debug "RepositoryPath:      $repositoryPath"
+Write-Debug "RepositoryNamespace: $repositoryNamespace"
 
 if (-not $DoNotCloneRepository) {
 
@@ -114,15 +144,23 @@ elseif (-not (Test-Path ".git" -PathType Container)) {
 
 if (-not $DoNotCopyTemplate) {
 
-    git clone --depth 1 "$($gitRepoPrefix)common/package-templates.git" "_template"
+    if (-not $debugUseLocalTemplates) {
 
-    Remove-Item "_template\.git" -Recurse -Force
-    Copy-Item "_template\*" "." -Recurse -Force
-    # Copy-Item "$PSScriptRoot\..\package-templates\*" "." -Recurse -Force
-    Remove-Item "_template" -Recurse -Force
+        git clone --depth 1 "$($gitRepoPrefix)common/package-templates.git" "_template"
+
+        Remove-Item "_template\.git" -Recurse -Force
+        Copy-Item "_template\*" "." -Recurse -Force
+        Remove-Item "_template" -Recurse -Force
+
+    } else {
+        Copy-Item "$PSScriptRoot\..\package-templates\*" "." -Recurse -Force
+    }
 
     if ((-not $DoNotRunTemplateScript) -and (Test-Path "Jumpstart.ps1")) {
-        & ".\Jumpstart.ps1" -RepositoryPath:$RepositoryPath -RepositoryName:$RepositoryName -RepositoryNamespace:$RepositoryNamespace -GitUseSsh:$gitUseSsh @RemainingArgumentsToPassToTemplate
+        Write-Host
+        & ".\Jumpstart.ps1" -VendorName:$vendorName -ApplicationName:$applicationName `
+            -RepositoryPath:$RepositoryPath -RepositoryNamespace:$RepositoryNamespace `
+            -GitUseSsh:$gitUseSsh @RemainingArgumentsToPassToTemplate
     }
 
 }
